pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "camel-society"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "camel-ai<0.2.24",
    "naptha-sdk @ file:///Users/vedantpadwal/Naptha/plugins/naptha-sdk",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
]

[tool.uv]
prerelease = "allow"

[tool.hatch.metadata]
allow-direct-references = true
